backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-multipart==0.0.6
pillow
sqlalchemy==2.0.9
psycopg2-binary==2.9.6
python-dotenv==1.0.0
torch
torchvision
matplotlib
sympy
mlflow
wandb
tqdm
numpy
scikit-learn
opencv-python
pydantic
nltk
